requirements.txt
pandas
requests
pyyaml
pytest
